% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{abj_new}
\alias{abj_new}
\title{Cria novo relatório com base em template.}
\usage{
abj_new(tipo = "pdf", nm = "relatorio.Rmd", path = "./")
}
\arguments{
\item{tipo}{por enquanto só pdf. Futuramente word e html?}

\item{nm}{nome do arquivo.}

\item{path}{caminho (pasta) do arquivo.}
}
\description{
Cria novo relatório com base em template da ABJ, com opcoes de tipo de doc.
}

