% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render_abj_pdf}
\alias{render_abj_pdf}
\title{Renderiza doc da ABJ.}
\usage{
render_abj_pdf(arq, toc = FALSE)
}
\arguments{
\item{arq}{nome do arquivo.}

\item{toc}{table of contents.}
}
\description{
Renderiza doc da ABJ usando \code{\link{pdf_document2}}.
}
